# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  geregetemplateai_internal_http_datatransfers_requests.OrgCreateRequest:
    properties:
      kind:
        enum:
        - ministry
        - agency
        - soe
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      parent_id:
        type: string
    required:
    - name
    - parent_id
    type: object
  geregetemplateai_internal_http_datatransfers_requests.OrgUpdateRequest:
    properties:
      kind:
        enum:
        - root
        - ministry
        - agency
        - soe
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  geregetemplateai_internal_http_datatransfers_requests.VoiceTranslateRequest:
    properties:
      audio_base64:
        type: string
      mime_type:
        description: |-
          MimeType-г allowlist-аар хязгаарлана — browser-ийн MediaRecorder-аас
          гардаг танигдсан багц л зөвшөөрнө (codecs параметрийг frontend хасдаг).
          Дурын мөрийг Gemini рүү дамжуулахгүй.
        enum:
        - audio/webm
        - audio/mp4
        - audio/ogg
        - audio/wav
        type: string
      source_lang:
        enum:
        - mn
        - en
        type: string
    required:
    - audio_base64
    - mime_type
    - source_lang
    type: object
...
  geregetemplateai_internal_http_datatransfers_responses.OrganizationResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      kind:
        type: string
      name:
        type: string
      parent_id:
        type: string
      path:
        type: string
      updated_at:
        type: string
    type: object
...
  geregetemplateai_internal_http_datatransfers_responses.VoiceTranslateResponse:
    properties:
      audio_base64:
        type: string
      audio_mime:
        type: string
      created_at:
        type: string
      id:
        type: string
      source_lang:
        type: string
      source_text:
        type: string
      target_lang:
        type: string
      translated_text:
        type: string
    type: object
...
  geregetemplateai_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Gerege Template AI v1.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: Fiber v3 + GORM (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; Gin -> Fiber v3 болон sqlx -> GORM руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Gerege Template AI v1.0 API
  version: "1.0"
paths:
...
  /organizations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.OrganizationResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Байгууллагын модыг жагсаах
      tags:
      - organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.OrgCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.OrganizationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Байгууллага үүсгэх (parent доор)
      tags:
      - organizations
  /organizations/{id}:
    delete:
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Байгууллага устгах (дэд мод бүхэлдээ)
      tags:
      - organizations
    put:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      - description: Organization
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.OrgUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.OrganizationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Байгууллагын нэр/төрөл засах
      tags:
      - organizations
...
  /voice/translate:
    post:
      consumes:
      - application/json
      description: Base64 аудио хэрчмийг хүлээн авч, бичвэрлэж (STT), нөгөө хэл рүү
        орчуулж, орчуулгыг яриа болгож (TTS) буцаана. Хариунд эх бичвэр, орчуулга
        болон тоглуулахад бэлэн WAV (base64) орно.
      parameters:
      - description: Voice clip
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.VoiceTranslateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.VoiceTranslateResponse'
              type: object
        "400":
          description: Malformed body / audio
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Unauthenticated
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Daily limit exceeded
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "503":
          description: Voice service not configured
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Дуу хоолойн орчуулга (MN↔EN)
      tags:
      - voice
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"